Feature/5359 test files group 15 (#1196)
* fix typo

* Update ExportTab.test.js

* Update Export.test.js

* Update ExportTab.test.mocks.js

* add data test id to export table

* mock props

* update tests

* standardized tests

-used reusable render
-moved mocks into constants/functions file

* remove extra wrapped act around render
