--- src/components/export/ExportTab/ExportTab.test.mocks.js
@@ -1,247 +1,0 @@
-const reportingPeriodId = 'reportingPeriodId'
-
-export const EXPORT_TAB_TEST_EXPORT_STATE = {
-  checkedDataTypes: [
-    'monitoring-plan'
-  ],
-  reportingPeriodId
-}
-
-export const getReportingPeriods = {
-  data: [
-    {
-      id: reportingPeriodId,
-      calendarYear: 1993,
-      quarter: 1,
-      beginDate: "1993-01-01",
-      endDate: "1993-03-31",
-      periodDescription: "1993 QTR 1",
-      periodAbbreviation: "1993 Q1",
-      archiveInd: 0,
-      selected: false,
-    },
-    {
-      id: 2,
-      calendarYear: 1993,
-      quarter: 2,
-      beginDate: "1993-04-01",
-      endDate: "1993-06-31",
-      periodDescription: "1993 QTR 2",
-      periodAbbreviation: "1993 Q2",
-      archiveInd: 0,
-      selected: false,
-    },
-  ],
-}
-
-export const mockReportingPeriod = [
-  {
-    id: reportingPeriodId,
-    calendarYear: 1993,
-    quarter: 1,
-    beginDate: "1993-01-01",
-    endDate: "1993-03-31",
-    periodDescription: "1993 QTR 1",
-    periodAbbreviation: "1993 Q1",
-    archiveInd: 0,
-    selected: false,
-  },
-  {
-    id: 2,
-    calendarYear: 1993,
-    quarter: 2,
-    beginDate: "1993-04-01",
-    endDate: "1993-06-31",
-    periodDescription: "1993 QTR 2",
-    periodAbbreviation: "1993 Q2",
-    archiveInd: 0,
-    selected: false,
-  },
-];
-
-export const exportQAResponse = {
-  orisCode: 3776,
-  testSummaryData: [
-    {
-      id: "TWCORNEL5-5438209079BE4E7C83507AFC1D8DA532",
-      locationId: "5",
-      stackPipeId: "CS0AAN",
-      unitId: null,
-      testTypeCode: "F2LCHK",
-      monitoringSystemID: "AA4",
-      componentID: null,
-      spanScaleCode: null,
-      testNumber: "201904010000AA4",
-      testReasonCode: "QA",
-      testDescription: null,
-      testResultCode: "FEW168H",
-      calculatedTestResultCode: "FEW168H",
-      beginDate: null,
-      beginHour: null,
-      beginMinute: null,
-      endDate: null,
-      endHour: null,
-      endMinute: null,
-      gracePeriodIndicator: null,
-      calculatedGracePeriodIndicator: null,
-      year: 2019,
-      quarter: 1,
-      testComment: null,
-      injectionProtocolCode: null,
-      calculatedSpanValue: null,
-      evalStatusCode: null,
-      userId: "bvick",
-      addDate: "4/24/2019, 5:27:06 PM",
-      updateDate: null,
-      reportPeriodId: 105,
-      calibrationInjectionData: [],
-      linearitySummaryData: [],
-    }
-  ],
-  certificationEventData: [
-    {
-      id: "string",
-      locationId: "string",
-      lastUpdated: "2023-02-17T20:37:37.814Z",
-      updatedStatusFlag: "string",
-      needsEvalFlag: "string",
-      checkSessionId: "string",
-      submissionId: 0,
-      submissionAvailabilityCode: "string",
-      pendingStatusCode: "string",
-      evalStatusCode: "string",
-      userId: "string",
-      addDate: "string",
-      updateDate: "string",
-      stackPipeId: "string",
-      unitId: "string",
-      monitoringSystemID: "string",
-      componentID: "string",
-      qaCertEventCode: "string",
-      qaCertEventDate: "2023-02-17T20:37:37.814Z",
-      qaCertEventHour: 0,
-      requiredTestCode: "string",
-      conditionalBeginDate: "2023-02-17T20:37:37.814Z",
-      conditionalBeginHour: 0,
-      completionTestDate: "2023-02-17T20:37:37.814Z",
-      completionTestHour: 0
-    }
-  ],
-  testExtensionExemptionData: [
-    {
-      stackPipeId: "string",
-      unitId: "string",
-      year: 0,
-      quarter: 0,
-      monitoringSystemID: "string",
-      componentID: "string",
-      spanScaleCode: "string",
-      id: "string",
-      locationId: "string",
-      reportPeriodId: 0,
-      checkSessionId: "string",
-      submissionId: "string",
-      submissionAvailabilityCode: "string",
-      pendingStatusCode: "string",
-      evalStatusCode: "string",
-      userId: "string",
-      addDate: "string",
-      updateDate: "string",
-      hoursUsed: 0,
-      fuelCode: "string",
-      extensionOrExemptionCode: "string"
-    }
-  ]
-}
-
-export const selectedConfig = {
-  id: "TWCORNEL5-C0E3879920A14159BAA98E03F1980A7A",
-  facId: 1,
-  configTypeCode: null,
-  lastUpdated: null,
-  updatedStatusFlag: "Y",
-  needsEvalFlag: null,
-  checkSessionId: null,
-  orisCode: 3,
-  name: "1, 2, CS0AAN",
-  beginReportPeriodId: 92,
-  endReportPeriodId: null,
-  active: true,
-  pendingStatusCode: null,
-  evalStatusCode: null,
-  unitStackConfigurations: [
-    {
-      id: "CAMD-18DD175CE7EF4256B78469865D84F576",
-      unitId: "1",
-      stackPipeId: "CS0AAN",
-      unitRecordId: 1,
-      stackPipeRecordId: "MDC-CCB8D6D0D4E34D24A99C01DCD14078DF",
-      beginDate: "1995-01-01",
-      endDate: null,
-      userId: "PQA09Q1",
-      addDate: "2009-02-20T14:57:04.000Z",
-      updateDate: null,
-      active: true,
-    },
-    {
-      id: "CAMD-BCB0E6D124C947418397A19FD7B4FB77",
-      unitId: "2",
-      stackPipeId: "CS0AAN",
-      unitRecordId: 2,
-      stackPipeRecordId: "MDC-CCB8D6D0D4E34D24A99C01DCD14078DF",
-      beginDate: "1995-01-01",
-      endDate: null,
-      userId: "PQA09Q1",
-      addDate: "2009-02-20T14:57:04.000Z",
-      updateDate: null,
-      active: true,
-    },
-  ],
-  locations: [
-    {
-      id: "6",
-      unitRecordId: 1,
-      unitId: "1",
-      stackPipeRecordId: null,
-      stackPipeId: null,
-      name: "1",
-      type: "unit",
-      active: true,
-      activeDate: null,
-      retireDate: null,
-      nonLoadBasedIndicator: 0,
-    },
-    {
-      id: "7",
-      unitRecordId: 2,
-      unitId: "2",
-      stackPipeRecordId: null,
-      stackPipeId: null,
-      name: "2",
-      type: "unit",
-      active: true,
-      activeDate: null,
-      retireDate: null,
-      nonLoadBasedIndicator: 0,
-    },
-    {
-      id: "5",
-      unitRecordId: null,
-      unitId: null,
-      stackPipeRecordId: "MDC-CCB8D6D0D4E34D24A99C01DCD14078DF",
-      stackPipeId: "CS0AAN",
-      name: "CS0AAN",
-      type: "stack",
-      active: true,
-      activeDate: "1995-01-01",
-      retireDate: null,
-      nonLoadBasedIndicator: null,
-    },
-  ],
-  userId: "bvick",
-  addDate: "2015-10-26T10:49:28.000Z",
-  updateDate: "2021-07-26T11:26:00.000Z",
-  submissionId: 1441000,
-  submissionAvailabilityCode: "UPDATED",
-  lastEvaluatedDate: "2022-04-25T13:59:00.000Z",
-};
