+++ src/components/export/ExportTab/ExportTablesContainer.js
@@ -106,7 +106,7 @@ export const ExportTablesContainer = ({
 
   const dataTable = (
     <div className="margin-x-3 margin-y-4" id={divId}>
-      <h4 className="margin-y-1">{tableTitle}</h4>
+      <h4 className="margin-y-1" data-testid="export-table-title">{tableTitle}</h4>
       <DataTable
         className="data-display-table maxh-mobile overflow-y-scroll fixed-table-header"
         responsive={true}
