+++ src/components/export/ExportTab/ExportTablesContainer.test.js
@@ -1,14 +1,10 @@
 import React from "react";
 import { render, screen } from "@testing-library/react";
-import ExportTablesContainer from "./ExportTablesContainer";
-import axios from "axios";
-import MockAdapter from "axios-mock-adapter";
-import { qaTestSummaryCols } from "../../../utils/constants/tableColumns";
-import { secureAxios } from "../../../utils/api/easeyAuthApi";
-const mock = new MockAdapter(axios);
 
-jest.mock("../../../utils/api/easeyAuthApi");
-secureAxios.mockImplementation((options) => axios(options));
+import * as qaCertificationApi from "../../../utils/api/qaCertificationsAPI";
+
+import ExportTablesContainer from "./ExportTablesContainer";
+import { getMockExportQa } from "../../../mocks/functions";
 
 const props = {
   selectionData: { beginDate: "1/1/11", endDate: "1/1/11" },
@@ -23,94 +19,22 @@ const props = {
   tableTitle: "Test Summary",
   dataKey: "testSummaryData",
 };
-const exportUrl =
-  "https://api.epa.gov/easey/dev/qa-certification-mgmt/export?facilityId=3776&beginDate=1/1/11&endDate=1/1/11";
-const response = {
-  orisCode: 3776,
-  testSummaryData: [
-    {
-      id: "TWCORNEL5-5438209079BE4E7C83507AFC1D8DA532",
-      locationId: "5",
-      stackPipeId: "CS0AAN",
-      unitId: null,
-      testTypeCode: "F2LCHK",
-      monitoringSystemID: "AA4",
-      componentID: null,
-      spanScaleCode: null,
-      testNumber: "201904010000AA4",
-      testReasonCode: "QA",
-      testDescription: null,
-      testResultCode: "FEW168H",
-      calculatedTestResultCode: "FEW168H",
-      beginDate: null,
-      beginHour: null,
-      beginMinute: null,
-      endDate: null,
-      endHour: null,
-      endMinute: null,
-      gracePeriodIndicator: null,
-      calculatedGracePeriodIndicator: null,
-      year: 2019,
-      quarter: 1,
-      testComment: null,
-      injectionProtocolCode: null,
-      calculatedSpanValue: null,
-      evalStatusCode: null,
-      userId: "bvick",
-      addDate: "4/24/2019, 5:27:06 PM",
-      updateDate: null,
-      reportPeriodId: 105,
-      calibrationInjectionData: [],
-      linearitySummaryData: [],
-      rataData: [],
-      flowToLoadReferenceData: [],
-      flowToLoadCheckData: [
-        {
-          id: "TWCORNEL5-6B67F4286CBD4867BC981E601A989626",
-          testSumId: "TWCORNEL5-5438209079BE4E7C83507AFC1D8DA532",
-          testBasisCode: "Q",
-          biasAdjustedIndicator: null,
-          avgAbsolutePercentDiff: null,
-          numberOfHours: null,
-          numberOfHoursExcludedForFuel: null,
-          numberOfHoursExcludedRamping: null,
-          numberOfHoursExcludedBypass: null,
-          numberOfHoursExcludedPreRATA: null,
-          numberOfHoursExcludedTest: null,
-          numberOfHoursExcMainBypass: null,
-          operatingLevelCode: "H",
-          userId: "bvick",
-          addDate: "4/24/2019, 5:27:07 PM",
-          updateDate: null,
-        },
-      ],
-      cycleTimeSummaryData: [],
-      onlineOfflineCalibrationData: [],
-      fuelFlowmeterAccuracyData: [],
-      transmitterTransducerData: [],
-      fuelFlowToLoadBaselineData: [],
-      appECorrelationTestSummaryData: [],
-      fuelFlowToLoadTestData: [],
-      unitDefaultTestData: [],
-      hgSummaryData: [],
-      testQualificationData: [],
-      protocolGasData: [],
-      airEmissionTestingData: [],
-    },
-  ],
-  certificationEventData: [],
-  testExtensionExemptionData: [],
-};
-mock.onGet(exportUrl).reply(200, response);
 
-test("renders QA Test Summary table", async () => {
-  jest.setTimeout(10000);
-  render(<ExportTablesContainer {...props} />);
-  const testSummaryTitle = await screen.findByText(/test summary/i);
-  expect(testSummaryTitle).toBeInTheDocument();
-  const columnheaders = await screen.findAllByRole("columnheader");
-  expect(columnheaders.length).toBe(qaTestSummaryCols.length);
-  expect(await screen.findByText("Unit/Stack Pipe ID")).toBeVisible();
-  expect(await screen.findByText("System/Component ID")).toBeVisible();
-  expect(await screen.findByText("Test Type Code")).toBeVisible();
+beforeEach(() => {
+  jest.spyOn(qaCertificationApi, "exportQA").mockResolvedValue({
+    data: getMockExportQa(),
+    status: 200
+  });
+})
+
+afterEach(() => {
+  jest.clearAllMocks();
+})
+
+test("renders ExportTablesContainer", async () => {
+  await render(
+    <ExportTablesContainer {...props} />
+  );
+  const tableTitle = await screen.findByTestId('export-table-title')
+  expect(tableTitle).toBeDefined();
 });
